expenses: address PR review feedback — auth handler, withdrawable statuses, IBAN normalize, job→service refactor
Fixes 6 review findings on PR #480 (the 7th — coordinator endorsement stub —
is tracked as follow-up nobodies-collective/Humans#696):

Auth + status invariants:
- ExpensesController.cs: withdrawableStatuses now matches the section invariant
  (Submitted, CoordinatorEndorsed, Approved — Draft removed, Approved added).
- ExpenseReportAuthorizationHandler.cs: Withdraw grant aligned to the same set.
  Per the actors table, FinanceAdmin (and by inclusion Admin) now gets the
  coordinator capabilities Endorse + CoordinatorReject (gated to Submitted),
  in addition to the finance-only operations. The two Admin/FinanceAdmin blocks
  are merged since IsFinanceAdmin already includes Admin.

ProfileService.SetIbanAsync: removed duplicate .Replace(" ", "") — both args
were U+0020; the second call was a no-op.

Job → service refactor (design-rules §2: jobs may not bypass the owning service):
- IExpenseReportService gains DrainHoldedOutboxAsync and PollHoldedPaidStatusAsync.
  ExpenseReportService injects IHoldedClient and owns the per-event drain loop
  + the SepaSent → Paid polling loop (moved verbatim from the jobs).
- HoldedExpenseOutboxJob and ExpensePaidPollingJob shrink to one-line Hangfire
  wrappers calling the service. Concurrent-execution attributes preserved.
- Test surface moved: full drain/poll coverage now lives at
  ExpenseReportServiceHoldedOutboxTests / ExpenseReportServiceHoldedPollingTests.
  Job test files keep small smoke tests verifying delegation + batch size.

Architecture baseline:
- NoObsoleteNavReads baseline drops the stale ExpenseReportAuthorizationHandler
  User#3 entry (removed when the duplicate isAdmin block was consolidated).

Auth handler tests inverted: the previous assertions pinned the buggy behavior
(submitter can't withdraw Approved; FinanceAdmin can't endorse) and are now
flipped to match the section invariant.
