+++ tests/Humans.Application.Tests/Architecture/Baselines/NoObsoleteNavReads.baseline.txt
@@ -198,7 +198,6 @@ src/Humans.Web/Authorization/MembershipRequiredFilter.cs:User#1
 src/Humans.Web/Authorization/Requirements/ActiveMemberOrShiftAccessHandler.cs:User#1
 src/Humans.Web/Authorization/Requirements/ExpenseReportAuthorizationHandler.cs:User#1
 src/Humans.Web/Authorization/Requirements/ExpenseReportAuthorizationHandler.cs:User#2
-src/Humans.Web/Authorization/Requirements/ExpenseReportAuthorizationHandler.cs:User#3
 src/Humans.Web/Authorization/Requirements/IbanAccessHandler.cs:User#1
 src/Humans.Web/Authorization/Requirements/IbanAccessHandler.cs:User#2
 src/Humans.Web/Authorization/Requirements/IbanAccessHandler.cs:User#3
